refacto OutputList (OutputsDetails) to split it
